fix: move test for EmissionsComparison and correct path to images
