+++ src/view/components/results/comparison/EmissionsComparison.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { CalculationData } from "../../../../data/calculations/ICalculationsRepository";
 import { getEmissionsComparison } from "../../../../utils/helpers/getEmissionComparison";
 
+import CoffeeCupPath from "../../../../assets/coffee-cup.svg";
+import GoogleSearchPath from "../../../../assets/google-search.svg";
+import PlasticBottlePath from "../../../../assets/plastic-bottle.svg";
 import { RealLifeExample } from "../../../../data/constants/RealLifeComparison";
-import CoffeeCupPath from "../../../assets/coffee-cup.svg";
-import GoogleSearchPath from "../../../assets/google-search.svg";
-import PlasticBottlePath from "../../../assets/plastic-bottle.svg";
 
 export const EmissionsComparison = ({
     calculation,
